Move USD.NET.Unity to the packa (USDU-137) (#240)
This moves the Unity specific part of the bindings (USD.NET.Unity) inside the package to allow faster iterations and better test environment.

The old copyright notices have been removed from the files and the unit/regression tests have been moved from the VS project to the package. Assembly defs have been updated accordingly.

+++ package/com.unity.formats.usd/Tests/Runtime/BaseFixture.cs
@@ -30,8 +30,13 @@ namespace Unity.Formats.USD.Tests
         {
             foreach (var file in m_filesToDelete)
             {
-                File.Delete(file);
+                if (File.Exists(file))
+                {
+                    File.Delete(file);
+                }
             }
+
+            m_filesToDelete.Clear();
         }
     }
 }
